creating visualize
